refactor(terminal): extract new session menu helpers
